# Speed up GCRS<->(TETE, CIRS) by smarter location conversions