Following the suggestion by @mkbrewer in #11067, this speeds up TETE<->GCRS considerably by using the `rbpn` matrix also for getting the sidereal time for the conversion of the observer location from ITRS to TETE.

fixes #11067

~Note: need to re-unify code so for now just a draft so we have something to look at.~ (done)

EDIT: quick timing:
```
from astropy.coordinates import SkyCoord, EarthLocation
from astropy import units as u

tete = SkyCoord(10*u.deg, 11*u.deg, frame='tete', location=EarthLocation(11*u.deg, 12*u.deg))
%timeit tete.transform_to('gcrs')
# current master
# 16.6 ms ± 1.13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# this PR
# 6.51 ms ± 609 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```