@mhvk 2020-11-20
Sorry, this is definitely not right: obsgeoloc and obsgeovel become all 0; I'm doing something stupid.