@mhvk 2020-11-20
More generally, a sequence of matrix multiplications cannot make the velocity non-zero, so `obsgeovel` is wrong even if I get `obsgeoloc` to be right.