@mkbrewer 2020-11-20
That check against Horizons is off by 288 mas. This suggests that there is something wrong at the polar motion level.