@mhvk 2020-11-20
OK, I get tests to pass again by reverting to what used to be done in `get_gcrs_posvel()` before we had velocities in our coordinates -- just calculating the velocity from the earth's rotation rate. So, this avoids the finite difference calls to get the velocity (which is why the speed-up is as large as it is: it not just halves the number of calls to `pnm06`, but quarters it!)