@mkbrewer 2020-11-20
Since the CIRS  and TETE transforms are now basically the same, I am curious whether or not `test_tete_transforms()` will work in this PR with the tighter tolerance of 1 mm. Would you mind trying that here?