@mkbrewer 2020-11-20
The ERFA routine is `pvtob()`, but it does the geodetic to geocentric conversion also which not necessary with an `EarthLocation`.