@mkbrewer 2020-11-20
Hmm, I wonder why my comment was automatically deleted there. Anyway, yes, we both thought of that at basically the same time.