@mhvk 2020-11-20
Ah, from that routine, it looks like I should be calculating the velocity in CIRS, as was done in #6699, so this needs an update.