@mkbrewer 2020-11-20
I think you are fine there. You are just calculating the velocity after the final rotation. Otherwise, you'd need to rotate the velocity also. See `apco()` for an example of this. In fact, we could optimize `ErfaAstrom.apco()` a bit by using your transform there also.