@mkbrewer 2020-11-20
btw, in fact, `pvtob()` works in either CIRS or TETE, see note 3) there.