@mkbrewer 2020-11-20
I don't see the point of making that change, unless you actually want the velocity in CIRS or TETE, you might as well calculate it in GCRS. That saves one extra rotation.