@mkbrewer 2020-11-20
Well, it is perpendicular to the CIP. The transformation to GCRS just aligns the CIP with the ICRF frame. That's why `apco()` rotates both the position and velocity into the GCRS. Remember, in that transformation we are going to of date to J2000.