@mhvk 2020-11-21
It's not quite perpendicular, though, and for some applications that small difference may matter (which was the reason for #6699 and the subsequent #6861)
```
from astropy import units as u
from astropy.time import Time
from astropy.coordinates import CIRS, GCRS
obstime = Time('J2010')
cirs = CIRS(x=10000*u.km, y=0.*u.km, z=0.*u.km, representation_type='cartesian', obstime=obstime)
cirs.transform_to(GCRS(obstime=obstime)).cartesian
# <CartesianRepresentation (x, y, z) in km
#    (9999.99496686, -0.00014031, -10.03308023)>
```