@mhvk 2020-11-21
My lack of being convinced is just because of @StuartLittlefair's old #6699...  But strengthened by the fact that with the current PR, `loc.get_gcrs_posvel()` and my quick routine give inconsistent velocities. If I apply the velocity in the CIRS or TETE frame, then they become consistent. I've just pushed a new commit that includes tests for this - do have a look, since if this is wrong, it affects people trying to get accurate radial velocities to hunt for planets...