@mhvk 2020-11-21
If we go this route, we should make sure `OMEGA_EARTH` is exactly consistent with the value used in ERFA's `pvtob` (unfortunately, not exported in their .h file). It might also be a good idea to rewrite `EarthLocation.get_gcrs_posvel()` to use the same method - I had not really realized how by just using the standard transformations, the velocity requires multiple evaluations. Though perhaps it is better to try to rewrite the transformation mechanism itself so that it uses knowledge of how velocities change.