@StuartLittlefair 2020-11-21
A lot happens when I’m asleep! It looks like you’ve sorted out all the issues; as you say velocity should be added in CIRS or TETE and then rotated from there.

I’m very pleased this also fixes the slight inaccuracies in the ICRS round trip as well as providing a massive speed up.