@mkbrewer 2020-11-21
There is a slight discrepancy between SOFA and Astropy regarding `OMEGA`. Using SOFA's definition of `D2PI`, I'm seeing:
```
OM = 7.292115146706980e-05
```