@mhvk 2020-11-21
@mkbrewer - thanks, I think we might as well be consistent with what SOFA has.

I'll make another commit that updates the value and adds a changelog entry about the speed and accuracy improvements.