@mhvk 2020-11-21
OK, I changed `OMEGA_EARTH` to match ERFA (and found where they got the value from...), and also added changelog and what's-new entries. @mkbrewer - for the latter, I mention you specifically, as I wanted to ensure your really helpful advice is recognized! (Normally, people are recognized by commits, but that process is not the only one...)